@@ -1,22 +1,10 @@
-import Vue from 'vue'
-import App from './App.vue'
-import vuetify from './plugins/vuetify';
-import VueRouter from 'vue-router'
-import { routes } from './routes';
-import store from './store/store';
+import { createApp } from 'vue';
 
-Vue.config.productionTip = false
+import App from './App.vue';
+import FriendContact from './components/FriendContact.vue';
 
-Vue.use(VueRouter);
+const app = createApp(App);
 
-const router = new VueRouter({
-  mode: 'history',
-  routes
-});
+app.component('friend-contact', FriendContact);
 
-new Vue({
-  vuetify,
-  router,
-  store,
-  render: h => h(App)
-}).$mount('#app')
+app.mount('#app');
